# [CK-Tile] Enable vectorized reads on all layouts & improve perf. 