## Proposed changes

This PR:

- enables vectorized reads on all layouts for input tensors {RR,RC,CC,CR} 
- Improve ck-tile universal gemm compute bound pipeline perf by from ~4% to 26%.

```sh
$ ./bin/tile_example_universal_gemm -a_layout=R -b_layout=R
Launching kernel with args: grid: {15, 16, 1}, blocks: {256, 1, 1}
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =2048 StrideB =4096 StrideC =4096 : 0.189726 ms, 339.567 TFlops, 337.135 GB/s,
$ ./bin/tile_example_universal_gemm -a_layout=R -b_layout=C
Launching kernel with args: grid: {15, 16, 1}, blocks: {256, 1, 1}
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =2048 StrideB =2048 StrideC =4096 : 0.187532 ms, 343.538 TFlops, 341.078 GB/s,
$ ./bin/tile_example_universal_gemm -a_layout=C -b_layout=C
Launching kernel with args: grid: {15, 16, 1}, blocks: {256, 1, 1}
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =3840 StrideB =2048 StrideC =4096 : 0.188303 ms, 342.133 TFlops, 339.683 GB/s,
$ ./bin/tile_example_universal_gemm -a_layout=C -b_layout=R
Launching kernel with args: grid: {15, 16, 1}, blocks: {256, 1, 1}
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =3840 StrideB =4096 StrideC =4096 : 0.202261 ms, 318.522 TFlops, 316.241 GB/s,
```

```sh
./bin/tile_example_universal_gemm -a_layout=R -b_layout=R
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =2048 StrideB =4096 StrideC =4096 : 0.21723 ms, 296.572 TFlops, 294.448 GB/s,
./bin/tile_example_universal_gemm -a_layout=R -b_layout=C
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =2048 StrideB =2048 StrideC =4096 : 0.19486 ms, 330.62 TFlops, 328.252 GB/s
./bin/tile_example_universal_gemm -a_layout=C -b_layout=C
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =3840 StrideB =2048 StrideC =4096 : 0.227269 ms, 283.473 TFlops, 281.443 GB/s
./bin/tile_example_universal_gemm -a_layout=C -b_layout=R
Run Gemm kernel with M =3840 N =4096 K =2048 StrideA =3840 StrideB =4096 StrideC =4096 : 0.254704 ms, 252.939 TFlops, 251.128 GB/s,
```

## Checklist

Please put an `x` into the boxes that apply. You can also fill these out after creating the PR. If you're not sure, please don't hesitate to ask.

- [ x] I have added tests relevant to the introduced functionality, and the unit tests are passing locally
- [ ] I have added the test to REGRESSION_TESTS list defined at the top of CMakeLists.txt in tests/CMakeLists.txt, **IF** the test takes more than 30 seconds to run.
- [x ] I have added inline documentation which enables the maintainers with understanding the motivation
- [ x] I have removed the stale documentation which is no longer relevant after this pull request
- [ ] (If this change is user-facing) I have added release notes which provide the end users with a brief summary of the improvement from this pull request
- [x ] I have run `clang-format` on all changed files
- [x ] Any dependent changes have been merged

## Discussion